teste

+++ src/main/java/br/com/nexfar/applicationtest/Comparable/ComparatorPrice.java
@@ -8,6 +8,6 @@ public class ComparatorPrice implements Comparator<ListingAndReview> {
 
     @Override
     public int compare(ListingAndReview o1, ListingAndReview o2) {
-        return Float.compare(o1.getPrice(), o2.getPrice());
+        return Double.compare(o1.getPrice(), o2.getPrice());
     }
 }
